Fixed position updating/jumping on menu open
